DRN  一大推东西，你们注意点啊，但没有修改你们的东西（员工表，考勤表，角色表，请假，出差

+++ src/main/java/com/hmm/travel/dao/TravelDao.java
@@ -7,9 +7,8 @@ import org.springframework.stereotype.Repository;
 import com.hmm.travel.entity.Travel;
 
 
-
 @Repository
 public interface TravelDao extends JpaSpecificationExecutor<Travel> ,
-								PagingAndSortingRepository<Travel, Integer> {
+								PagingAndSortingRepository<Travel, Long> {
 	
 }
